change where feed is called
