+++ feed/index.html
@@ -7,7 +7,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <script src="https://kit.fontawesome.com/54f19440a6.js" crossorigin="anonymous"></script>
   <link rel="stylesheet" href="../css/style.css"/>
-  <script type="module" src="feed.js"></script>
+  <script type="module" src="../js/index.js"></script>
 </head>
 <body class="bg-gradient-to-b from-violet-300 to-orange-50 min-h-screen grid grid-rows-[auto_1fr_auto]">
   <header class="w-100 h-20 flex justify-between">
@@ -118,11 +118,9 @@
               <label for="text" class="pb-1 text-xl md:text-2xl">What is on your mind?</label>
               <textarea 
                 id="text"
-                name="text"  
-                rows="1" 
+                name="text"
                 placeholder="What is on your mind?" 
-                class="text-lg py-1 pl-5 outline shadow-md rounded-xl md:text-xl h-20">
-              </textarea>
+                class="text-lg py-1 pl-5 outline shadow-md rounded-xl md:text-xl h-20"></textarea>
             </div>
             <button class="bg-black text-white rounded-full px-10 py-2 text-xl hover:bg-slate-800">Post</button>
           </form>
