+++ js/ui/auth/sendPostToApi.js
@@ -1,0 +1,3 @@
+export function sendPostToApi(){
+  
+}
