+++ js/ui/posts/createOwnPosts.js
@@ -1,0 +1,3 @@
+export function createOwnPost (){
+  
+}
